refactor: common 모듈 규칙을 위한 리팩토링
