+++ lovebird-api/src/main/kotlin/com/lovebird/api/provider/AesEncryptor.kt
@@ -1,4 +1,4 @@
-package com.lovebird.common.utils
+package com.lovebird.api.provider
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
