+++ lovebird-api/src/main/kotlin/com/lovebird/api/service/external/PresignedUrlService.kt
@@ -4,7 +4,7 @@ import com.lovebird.api.dto.param.external.DiaryUploadPresignedUrlParam
 import com.lovebird.api.dto.param.external.ProfileUploadPresignedUrlParam
 import com.lovebird.api.dto.response.external.PresignedUrlListResponse
 import com.lovebird.api.dto.response.external.PresignedUrlResponse
-import com.lovebird.common.utils.FilenameFormatter
+import com.lovebird.api.provider.FilenameFormatter
 import com.lovebird.s3.provider.PresignedUrlProvider
 import org.springframework.stereotype.Service
