+++ lovebird-domain/src/main/kotlin/com/lovebird/domain/repository/reader/CoupleCodeReader.kt
@@ -1,6 +1,6 @@
 package com.lovebird.domain.repository.reader
 
-import com.lovebird.common.annotation.Reader
+import com.lovebird.domain.annotation.Reader
 import com.lovebird.common.enums.ReturnCode
 import com.lovebird.common.exception.LbException
 import com.lovebird.domain.entity.CoupleCode
