+++ lovebird-domain/src/main/kotlin/com/lovebird/domain/repository/writer/DiaryWriter.kt
@@ -1,6 +1,6 @@
 package com.lovebird.domain.repository.writer
 
-import com.lovebird.common.annotation.Writer
+import com.lovebird.domain.annotation.Writer
 import com.lovebird.domain.dto.command.DiaryUpdateRequestParam
 import com.lovebird.domain.entity.Diary
 import com.lovebird.domain.repository.jpa.DiaryJpaRepository
@@ -10,8 +10,8 @@ class DiaryWriter(
 	private val diaryJpaRepository: DiaryJpaRepository
 ) {
 
-	fun save(diary: Diary) {
-		diaryJpaRepository.save(diary)
+	fun save(diary: Diary): Diary {
+		return diaryJpaRepository.save(diary)
 	}
 
 	fun update(diary: Diary, param: DiaryUpdateRequestParam) {
